refactored DAO files to the correct location and standardized names
