+++ src/main/java/user/createinternship/CreateInternshipConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import service.dao.IUserDAO;
-import service.InternshipDBGateway;
+import service.dao.IInternshipDAO;
 
 @Configuration
 @ComponentScan("service")
@@ -13,7 +13,7 @@ import service.InternshipDBGateway;
 public class CreateInternshipConfig {
     @Bean
     @Primary
-    public CreateInternshipInputBoundary createInternshipInput(InternshipDBGateway internshipDataAccess,
+    public CreateInternshipInputBoundary createInternshipInput(IInternshipDAO internshipDataAccess,
                                                                IUserDAO userDataAccess){
         return new CreateInternshipInteractor(internshipDataAccess, userDataAccess);
     }
