+++ src/main/java/user/createinternship/CreateInternshipInputBoundary.java
@@ -3,6 +3,5 @@ package user.createinternship;
 public interface CreateInternshipInputBoundary {
 
     CreateInternshipResponseDS createInternship(CreateInternshipInputDS inputDS);
-
-    void test_update_review();
+    
 }
