+++ src/main/java/user/createinternship/CreateInternshipInteractor.java
@@ -5,7 +5,7 @@ import entity.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import service.dao.IUserDAO;
-import service.InternshipDBGateway;
+import service.dao.IInternshipDAO;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -14,12 +14,12 @@ import java.util.List;
 public class CreateInternshipInteractor implements CreateInternshipInputBoundary {
 
     @Autowired
-    final InternshipDBGateway internshipDataAccess;
+    final IInternshipDAO internshipDataAccess;
 
     @Autowired
     final IUserDAO userDataAccess;
 
-    public CreateInternshipInteractor(InternshipDBGateway internshipDataAccess, IUserDAO userDataAccess){
+    public CreateInternshipInteractor(IInternshipDAO internshipDataAccess, IUserDAO userDataAccess){
         this.internshipDataAccess = internshipDataAccess;
         this.userDataAccess = userDataAccess;
     }
@@ -32,8 +32,8 @@ public class CreateInternshipInteractor implements CreateInternshipInputBoundary
             User creator = userDataAccess.getUser(inputDS.getCreatorUsername());
             System.out.println(creator.getUsername());
             System.out.println(creator.getAccessLevel());
-            if (creator.getAccessLevel() == 0){
-                // if access level is 0, return failure.
+            if (!creator.getCorporateRep()){
+                // if user is not a corporate rep, return failure.
                 return new CreateInternshipResponseDS("Not authorized to create internship");
             }
             // create a new internship
@@ -48,28 +48,4 @@ public class CreateInternshipInteractor implements CreateInternshipInputBoundary
         }
     }
 
-    public void test_get_internship(){
-        try{
-            Internship internship = internshipDataAccess.getInternshipByID(1);
-            System.out.println(internship.getReviews());
-
-        } catch (Exception e){
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public void test_update_review(){
-        try{
-            List<Integer> integerList = new ArrayList<Integer>();
-            Internship internship = new Internship(3, integerList, "plumber", "corporaterep@mail.utoronto.ca");
-
-            internshipDataAccess.updateInternship(1, internship);
-
-            List<Internship> internships = internshipDataAccess.getInternshipsByCompany(3);
-            System.out.println(internships);
-
-        } catch (Exception e){
-            System.out.println(e.getMessage());
-        }
-    }
 }
