+++ src/test/java/user/postreview/PostReviewTest.java
@@ -9,8 +9,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.junit4.SpringRunner;
 import service.ServerStatus;
-import service.dao.IInternshipDAO;
 import service.dao.IReviewDAO;
+import service.dao.IInternshipDAO;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @RunWith(SpringRunner.class)
